requirements.txt
alpha_vantage==2.1.2
paho-mqtt==1.5.0
